+++ tests/it/main.rs
@@ -123,6 +123,7 @@ mod temporal_async;
 mod temporal_carry;
 mod temporal_every;
 mod true_singleton;
+mod type_decls;
 mod type_entity_xlang;
 mod type_graph;
 mod type_graph_ts;